Correctly handle session close
This commit corrects an issue where closing the session and setting it to `NULL` before the "end" session event fired caused the native `pcap2` module to segfault by referencing a NULL pointer (the session).

This is resolved by placing the session closing on the session end event emitter handler.  This way the session will only be cleaned up when it's done and we avoid the race condition.