Added IP address of the responding host to the Ping result
Althought we should know what IP address responds to us, there's little reason to leave this data out.  In this commit we add the IP address to the Ping response so we can always know what IP address is correlated with a resulting MAC address.  This could be helpful when scanning many IP adrdesses.